urdu_tts/requirements.txt
Django==1.11.4
django-ipware==1.1.6
djangorestframework==3.7.7
pytz==2017.2
wheel==0.24.0
python-dateutil==2.6.1
Fabric==1.14.0
gunicorn==19.7.1
whitenoise==3.3.0
